main page reordering
