+++ wo-scheduler-v3.sql
@@ -583,7 +583,8 @@ INSERT INTO user_bodyweight (app_user_id, weight, date)
     (1, 76.02208,'2024-03-22T07:34:12'),
     (1, 74.02627,'2024-04-02T07:34:12'),
     (1, 75.38705,'2024-04-05T07:34:12'),
-    (1, 74.6613,'2024-04-05T07:34:12');
+    (1, 74.6613,'2024-04-08T07:34:12'),
+    (1, 73.84484,'2024-04-10T07:34:12');
 
 INSERT INTO workout_days (workout_id, day)
     VALUES
